add appframe comp
